+++ motionEstimation/PointCloudVis.h
@@ -37,11 +37,11 @@ inline float* Eigen2float6(Eigen::Vector3f v, Eigen::Vector3f rgb) { static floa
 inline Eigen::Matrix<float,6,1> Eigen2Eigen(Eigen::Vector3f v, Eigen::Vector3f rgb) { return (Eigen::Matrix<float,6,1>() << v[0],v[1],v[2],rgb[0],rgb[1],rgb[2]).finished(); }
 inline std::vector<Eigen::Matrix<float,6,1> > AsVector(const Eigen::Matrix<float,6,1>& p1, const Eigen::Matrix<float,6,1>& p2) { 	std::vector<Eigen::Matrix<float,6,1> > v(2); v[0] = p1; v[1] = p2; return v; }
 
-void addCameraToVisualizer(const Eigen::Matrix3f& R, const Eigen::Vector3f& _t, float r, float g, float b, double s = 0.01 /*downscale factor*/, const std::string& name = "");
+void addCameraToVisualizer(const Eigen::Matrix3f& R, const Eigen::Vector3f& _t, float r, float g, float b, float s = 0.01 /*downscale factor*/, const std::string& name = "");
 void addCameraToVisualizer(const float R[9], const float t[3], float r, float g, float b);
-void addCameraToVisualizer(const float R[9], const float t[3], float r, float g, float b, double s);
-void addCameraToVisualizer(const cv::Vec3f& T, const cv::Matx33f& R, float r, float g, float b, double s, const std::string& name);
-void addCameraToVisualizer(const cv::Mat &currentPos, const cv::Mat& T, const cv::Mat &R, float r, float g, float b, double s, const std::string& name);
+void addCameraToVisualizer(const float R[9], const float t[3], float r, float g, float b, float s);
+void addCameraToVisualizer(const cv::Vec3f& T, const cv::Matx33f& R, float r, float g, float b, float s, const std::string& name);
+void addCameraToVisualizer(const cv::Mat &currentPos, const cv::Mat& T, const cv::Mat &R, float r, float g, float b, float s, const std::string& name);
 
 void AddLineToVisualizer(const std::vector<cv::Point3f>& pointCloud_1, const std::vector<cv::Point3f>& pointCloud_2, std::string name, const cv::Scalar &color);
